fix: correct type annotations for physics quantities from int to float
Parameters representing axis limits, momentum cuts, and IP cuts were
annotated as int because their defaults are integer literals, but they
accept float values semantically.

+++ macro/eventDisplay.py
@@ -139,7 +139,7 @@ def printMCTrack(n: int, MCTrack) -> None:
     )
 
 
-def dump(pcut: int = 0) -> None:
+def dump(pcut: float = 0) -> None:
     print("   #         pid   px    py      pz     vx      vy       vz      mid")
     n = -1
     for mcp in sTree.MCTrack:
